github-mcp/pyproject.toml
[project]
name = "github-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.27.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]
